+++ crates/executor/src/rlm_tests.rs
@@ -8,24 +8,16 @@ use crate::{
 use anyhow::{anyhow, bail};
 use async_trait::async_trait;
 use exoharness::{
-    BasicExoHarness, BasicExoHarnessConfig, Binding, EventData, EventQuery, EventQueryDirection,
-    ExoHarness, PutSecretRequest, SandboxBackendChoice, Secret, SecretBackendChoice, ToolRequest,
-    Uuid7,
+    BasicExoHarness, Binding, EventData, EventQuery, EventQueryDirection, ExoHarness,
+    PutSecretRequest, Secret, ToolRequest, Uuid7,
 };
-
-fn local_test_config(root: impl Into<std::path::PathBuf>) -> BasicExoHarnessConfig {
-    BasicExoHarnessConfig {
-        root: root.into(),
-        secret_backend: SecretBackendChoice::Static([7u8; 32]),
-        sandbox_backend: SandboxBackendChoice::LocalProcess,
-    }
-}
 use lingua::universal::{AssistantContent, UserContent};
 use lingua::{Message, UniversalStreamChunk};
 use serde_json::{Map, Value};
 use tempfile::TempDir;
 use tokio_stream::StreamExt;
 
+use crate::test_support::local_test_config;
 use crate::{CreateAgentRequest, CreateConversationRequest, Harness, RlmHarness};
 
 #[tokio::test(flavor = "current_thread")]
