+++ crates/exoharness/src/http_tests.rs
@@ -4,20 +4,13 @@ use std::sync::Arc;
 use futures::io::AsyncReadExt;
 use tempfile::TempDir;
 
+use crate::test_support::local_test_config;
 use crate::{
-    BasicExoHarness, BasicExoHarnessConfig, CreateSandboxRequest, ExoHarness, HttpExoHarness,
-    RunInSandboxRequest, SandboxBackendChoice, SandboxProcessEvent, SandboxProcessEventQuery,
-    SandboxProcessStatus, SandboxProcessStdin, SecretBackendChoice, StartSandboxProcessRequest,
-    WaitSandboxProcessRequest, WriteSandboxProcessInputRequest, serve_exoharness_http_listener,
+    BasicExoHarness, CreateSandboxRequest, ExoHarness, HttpExoHarness, RunInSandboxRequest,
+    SandboxProcessEvent, SandboxProcessEventQuery, SandboxProcessStatus, SandboxProcessStdin,
+    StartSandboxProcessRequest, WaitSandboxProcessRequest, WriteSandboxProcessInputRequest,
+    serve_exoharness_http_listener,
 };
-
-fn local_test_config(root: impl Into<std::path::PathBuf>) -> BasicExoHarnessConfig {
-    BasicExoHarnessConfig {
-        root: root.into(),
-        secret_backend: SecretBackendChoice::Static([7u8; 32]),
-        sandbox_backend: SandboxBackendChoice::LocalProcess,
-    }
-}
 
 struct HttpHarnessFixture {
     harness: Arc<dyn ExoHarness>,
