+++ crates/exoharness/src/lib.rs
@@ -22,6 +22,8 @@ mod secrets;
 pub mod server;
 #[cfg(all(not(target_arch = "wasm32"), feature = "basic-backend"))]
 mod storage;
+#[cfg(all(test, not(target_arch = "wasm32"), feature = "basic-backend"))]
+mod test_support;
 mod types;
 mod uuid7;
