+++ crates/exoharness/src/test_support.rs
@@ -1,0 +1,11 @@
+use std::path::PathBuf;
+
+use crate::{BasicExoHarnessConfig, SandboxBackendChoice, SecretBackendChoice};
+
+pub(crate) fn local_test_config(root: impl Into<PathBuf>) -> BasicExoHarnessConfig {
+    BasicExoHarnessConfig {
+        root: root.into(),
+        secret_backend: SecretBackendChoice::Static([7u8; 32]),
+        sandbox_backend: SandboxBackendChoice::LocalProcess,
+    }
+}
