+++ crates/exoharness/src/types.rs
@@ -281,29 +281,6 @@ pub enum EventData {
         sandbox_id: SandboxId,
         snapshot_id: SnapshotId,
     },
-    SandboxProcessStarted {
-        sandbox_id: SandboxId,
-        process_id: SandboxProcessId,
-        command: Vec<String>,
-        cwd: Option<String>,
-        mode: SandboxProcessMode,
-        stdin: SandboxProcessStdin,
-        output: SandboxProcessOutput,
-        lifecycle: SandboxProcessLifecycle,
-        status: SandboxProcessStatus,
-        provider_state: Option<Value>,
-    },
-    SandboxProcessStateUpdated {
-        sandbox_id: SandboxId,
-        process_id: SandboxProcessId,
-        status: SandboxProcessStatus,
-        provider_state: Option<Value>,
-    },
-    SandboxProcessEvent {
-        sandbox_id: SandboxId,
-        process_id: SandboxProcessId,
-        event: SandboxProcessEvent,
-    },
     Custom {
         event_type: String,
         payload: Value,
